requirements.txt
# YouTube
yt-dlp>=2024.1.0

# Speech-to-Text (Telugu)
openai-whisper>=20231117

# Local LLM
ollama>=0.4.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
pydantic>=2.0.0
httpx>=0.25.0

# Audio processing (required by Whisper)
ffmpeg-python>=0.2.0
